Revert "Move common code to parachain utils"
This reverts commit d58208b4f42f573504319a3645cc380e9762735a.
